+++ lib/Atlas/src/Functions.h
@@ -2,8 +2,8 @@
 
 #include <functional>
 
-typedef std::function<double(double, double, double)> Integration;
+typedef std::function<float(float, float, float)> Integration;
 
-static Integration LINEAR_INTEGRATION = [](double a, double b, double i) { return a + ((b - a) * i) ; };
-static Integration FLOOR_INTEGRATION = [](double a, double b, double i) { return a; };
-static Integration CEILING_INTEGRATION = [](double a, double b, double i) { return b; };
+static Integration LINEAR_INTEGRATION = [](float a, float b, float i) { return a + ((b - a) * i) ; };
+static Integration FLOOR_INTEGRATION = [](float a, float b, float i) { return a; };
+static Integration CEILING_INTEGRATION = [](float a, float b, float i) { return b; };
